+++ src/main/java/com/github/apehum/bundlecounter/mixin/GuiGraphicsMixin.java
@@ -1,29 +1,33 @@
 package com.github.apehum.bundlecounter.mixin;
 
 import com.github.apehum.bundlecounter.BundleCounter;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.component.BundleContents;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Redirect;
 
 @Mixin(GuiGraphics.class)
 public abstract class GuiGraphicsMixin {
-    @Redirect(method = "renderItemCount", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/item/ItemStack;getCount()I"))
-    private int getCount(ItemStack instance) {
+    @WrapOperation(
+            method = "renderItemCount",
+            at = @At(value = "INVOKE", target = "Lnet/minecraft/world/item/ItemStack;getCount()I")
+    )
+    private int getCount(ItemStack instance, Operation<Integer> original) {
         if (!BundleCounter.CONFIG.shouldRenderOnItem) {
-            return instance.getCount();
+            return original.call(instance);
         }
 
         if (instance.getCount() != 1) {
-            return instance.getCount();
+            return original.call(instance);
         }
 
         BundleContents contents = instance.get(DataComponents.BUNDLE_CONTENTS);
         if (contents == null) {
-            return instance.getCount();
+            return original.call(instance);
         }
 
         var weight = contents.weight();